Tint works! for the most part...
I now have a working tint.  It's functional, however it does not
necessarily work on black and white.  White stays white, and black turns
the color of the tint. I will work on this problem for the future, but I
think what I have right now is sufficient for the present.